+++ src/main/java/com/villagercycle/mixin/MerchantEntityAccessor.java
@@ -1,11 +1,12 @@
 package com.villagercycle.mixin;
 
 import net.minecraft.entity.passive.MerchantEntity;
+import net.minecraft.server.world.ServerWorld;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Invoker;
 
 @Mixin(MerchantEntity.class)
 public interface MerchantEntityAccessor {
 	@Invoker("fillRecipes")
-	void invokeFillRecipes();
+	void invokeFillRecipes(ServerWorld world);
 }